==> 00.change2counts.ipynb <==
import os,re,sys
import scanpy as sc
os.chdir('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/01.h5ad_counts')

frog=sc.read('./frog_counts_pr.h5ad')

dogshark=sc.read('./dogshark_counts_pr.h5ad')

bichir=sc.read('./bichir_counts_pr.h5ad')

lancelet=sc.read('./lancelet_counts.h5ad')

lungfish = sc.read('./lungfish_counts_pr.h5ad')

lamprey = sc.read('./lamprey_counts_pr.h5ad')

human = sc.read('./human_counts_pr.h5ad')

lancelet.obs.info()

adata=lungfish.copy()

adata.obs=adata.obs.astype('str')

adata.obs.info()

info = ['sample','cluster','species','cell type']

def dropobs(adata,obs,sp=''):
    for i in adata.obs.columns:
        if i not in obs:
            del adata.obs[i]
        else:
            adata.obs[i] = adata.obs[i].astype('category')
    del adata.var
    del adata.varm
    del adata.obsm
    del adata.uns
    del adata.obsp
    print(adata)
    adata.write(f'H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/01.h5ad_counts/{sp}_counts.h5ad')
    return None

dropobs(human,info,sp='human')

lancelet.obs=lancelet.obs.astype('str')
lancelet.write('./lancelet_counts_pr.h5ad')

frog


lancelet.var_names = ['LOC'+ x if '1184' in x else x for x in lancelet.var_names ]

lancelet.var_names

lancelet.write('./lancelet_counts.h5ad')
