==> 02.samap.ipynb <==
from samap.mapping import SAMAP
from samap.analysis import (get_mapping_scores, GenePairFinder,
                            sankey_plot, chord_plot, CellTypeTriangles, 
                            ParalogSubstitutions, FunctionalEnrichment,
                            convert_eggnog_to_homologs, GeneTriangles)
from samalg import SAM
import pandas as pd
import os,re,sys

# os.chdir('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap')
# fn1 = './01.h5ad_counts/bichir_counts.h5ad'
# fn2 = './01.h5ad_counts/dogshark_counts.h5ad'
# fn3 = './01.h5ad_counts/frog_counts.h5ad'
# fn4 = './01.h5ad_counts/human_counts.h5ad'
# fn5 = './01.h5ad_counts/lamprey_counts.h5ad'
# fn6 = './01.h5ad_counts/lancelet_counts.h5ad'
# fn7 = './01.h5ad_counts/lungfish_counts.h5ad'
# resolution={'hu':0.5,'fr':0.5,'lm':0.5,'lu':0.5,'do':0.5,'la':0.5,'bi':0.5}

# filenames = {'bi':fn1,'do':fn2,'fr':fn3,'hu':fn4,'lm':fn5,'la':fn6,'lu':fn7}
# # filenames = {'bi':fn1,'do':fn2}
# sm = SAMAP(filenames,f_maps = 'H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/maps/',save_processed=True,resolutions=resolution)
# # sm.run(pairwise=True,N_GENE_CHUNKS=5)
fn4 = './01.h5ad_counts/human_counts_pr.h5ad'
# fn5 = './01.h5ad_counts/lamprey_counts_pr.h5ad'
# fn2 = './01.h5ad_counts/dogshark_counts_pr.h5ad'
fn7 = './01.h5ad_counts/lungfish_counts_pr.h5ad'
# fn1 = './01.h5ad_counts/bichir_counts_pr.h5ad'
# fn3 = './01.h5ad_counts/frog_counts_pr.h5ad'
# fn6 = './01.h5ad_counts/lancelet_counts_pr.h5ad'
sam1=SAM()
sam1.load_data(fn4)

sam2=SAM()
sam2.load_data(fn7)

# sam3 = SAM()
# sam3.load_data(fn2)

# sam4 = SAM()
# sam4.load_data(fn7)

# sam5 = SAM()
# sam5.load_data(fn1)

# sam6 = SAM()
# sam6.load_data(fn3)

# sam7 = SAM()
# sam7.load_data(fn6)

# os.chdir('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap')
# fn1 = './01.h5ad_counts/bichir_counts_pr.h5ad'
# fn2 = './01.h5ad_counts/dogshark_counts_pr.h5ad'
# fn3 = './01.h5ad_counts/frog_counts_pr.h5ad'
# fn4 = './01.h5ad_counts/human_counts_pr.h5ad'
# fn5 = './01.h5ad_counts/lamprey_counts_pr.h5ad'
# fn6 = './01.h5ad_counts/lancelet_counts_pr.h5ad'
# fn7 = './01.h5ad_counts/lungfish_counts_pr.h5ad'
# sam1=SAM()
# sam1.load_data(fn1)

# sam2=SAM()
# sam2.load_data(fn2)

# sam3=SAM()
# sam3.load_data(fn3)

# sam4=SAM()
# sam4.load_data(fn4)

# sam5=SAM()
# sam5.load_data(fn5)

# sam6=SAM()
# sam6.load_data(fn6)

# sam7=SAM()
# sam7.load_data(fn7)

# sams = {'bi':sam1,'do':sam2,'fr':sam3, 'hu':sam4,'lm':sam5, 'la':sam6, 'lu':sam7}
sams = {'hu':sam1,'lu':sam2,}
sm = SAMAP(
        sams,
        f_maps = './maps/',
    )

sm.run(N_GENE_CHUNKS=10,crossK=20)  ##pairwise=True

smp = sm.samap


smp.adata.varp["homology_graph_reweighted"]

from scipy import sparse
sparse.save_npz('./homology_graph_test.npz', smp.adata.varp["homology_graph_reweighted"])
csr_matrix_variable = sparse.load_npz('./homology_graph_test.npz') 

csr_matrix_variable = sparse.load_npz('./homology_graph_test.npz') 

sm.plot_expression_overlap({'hu':'FGG'})

samap.adata

import scanpy as sc
import scanpy.external as sce
adata=samap.adata.copy()

sc.pl.umap(adata,color=["hu_CDH5","hu_ENG","hu_FLT1","hu_CLEC4G","hu_RAMP2"])

human = pd.read_csv('02.eggNOG/human.annotations.tsv',sep='\t',index_col=0)
lungfish = pd.read_csv('02.eggNOG/lungfish.annotations.tsv',sep='\t',index_col=0)

EGGs={'lu':lungfish,'hu':human}

keys = {'lu':'cluster','bi':'cluster'}

ortholog_pairs = convert_eggnog_to_homologs(sm,EGGs,og_key='eggNOG_OGs',taxon=7711) #33213 is the taxon code for Bilaterians

paralog_pairs = convert_eggnog_to_homologs(sm,EGGs,og_key='eggNOG_OGs',taxon=7742) #7742 is the taxon code for Vertebrate 

paralog_pairs

Psubs = ParalogSubstitutions(sm,ortholog_pairs,psub_thr=0.3)

Psubs.head(30)

import numpy as np
ortholog_pairs = np.sort(ortholog_pairs,axis=1)

human = pd.read_csv('02.eggNOG/human.annotations.tsv',sep='\t',index_col=0)

import re
human.index = [re.sub(r'\.','_',x) for x in human.index.to_list()]

human = pd.read_csv('02.eggNOG/human.annotations.tsv',sep='\t',index_col=0)
lungfish = pd.read_csv('02.eggNOG/lungfish.annotations.tsv',sep='\t',index_col=0)

EGGs={'lu':lungfish,'hu':human}
ortholog_pairs = convert_eggnog_to_homologs(sm,EGGs,og_key='eggNOG_OGs',taxon=7742) #7711 is the taxon code for Chordata
paralog_pairs = convert_eggnog_to_homologs(sm,EGGs,og_key='eggNOG_OGs',taxon=7711) #7742 is the taxon code for Vertebrate
# Psubs = ParalogSubstitutions(sm,ortholog_pairs,paralog_pairs=paralog_pairs,psub_thr=0.1)
