==> 06.Eggnog_Trans.ipynb <==
#!/usr/bin/env python
# -*- encoding: utf-8 -*-
'''
@文件        :Untitled-1.ipynb
@说明        :
@时间        :2023/02/20 14:27:40
@作者        :Wu Baosheng
@版本        :1.0
'''

import pandas as pd
import numpy as np
import re,sys,os
os.chdir('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap')

x=pd.read_csv('hydra.tsv',sep='\t',header=0,index_col=0)

x['eggNOG_OGs']

df = pd.read_csv('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/02.eggNOG/MM_k5p15yf6.emapper.annotations_lamprey.tsv',sep='\t',comment='#',header=None,index_col=0)

df.index.name=None

egg = []
for i in df[4]:
    tmp = i.split(',')
    x = ''
    for i in tmp:
        j = re.sub(r'\|.*','',i)
        x = j+"," + x
    egg.append(x)

df['eggNOG_OGs'] = egg

df.rename(columns={6:'KOG'},inplace=True)

df.to_csv('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/02.eggNOG/lamprey.annotations.tsv',sep='\t')

df.shape

df = pd.read_csv('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/02.eggNOG/human.annotations.tsv',sep='\t',header=0,index_col=0)

df = df[~df.index.duplicated(keep='first')]

df.sort_index(inplace=True)

df.index.is_unique

[print(x) for x in df.index if df.index.to_list().count(x)>1]

df
