==> 08.ortherlog_paralog_subtitution_evernts.ipynb <==
#!/usr/bin/env python
# -*- encoding: utf-8 -*-
'''
@文件        :08.ortherlog_paralog_pairs.ipynb
@说明        :
@时间        :2023/02/23 11:00:31
@作者        :Wu Baosheng
@版本        :1.0
'''

import os,re,sys
import pandas as pd
os.chdir('H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/05.subtitution_events')

df = pd.read_csv('lancelet_lamprey_dogshark.sub.txt',header=0,sep='\t')

df=df[df['la/lm subbed'] !='N.S.']

df

def count_sub(x):
    b = []
    a = x.split(';')
    for i in a:
        b.append(i[:2])
    return b.count('do')

df_lungfish_sub = pd.DataFrame()

df_lungfish_sub = pd.DataFrame(df['la/lm subbed'].apply(count_sub))

df_lungfish_sub

df_lungfish_sub.columns = ['lm_sub']

df_lungfish_sub['la_sub']=df['la/do subbed'].apply(count_sub)

df_lungfish_sub=df_lungfish_sub[df_lungfish_sub['lm_sub']>df_lungfish_sub['la_sub']]

df_lungfish_sub

df_lungfish_sub2=pd.DataFrame()

df_lungfish_sub2=pd.DataFrame(df['la/do subbed'].apply(count_sub))
df_lungfish_sub2.columns = ['la_sub']

df_lungfish_sub2['do_sub']=df['la/do subbed'].apply(count_sub)

df_lungfish_sub2=df_lungfish_sub2[df_lungfish_sub2['do_sub']>df_lungfish_sub2['la_sub']]

df_lungfish_sub2

sizu = list(set(df_lungfish_sub2.index).intersection(df_lungfish_sub.index))

df[df.index.isin(df_lungfish_sub2.index.to_list())].to_csv('lancelet_dogshark_sub.txt',sep='\t')

df[df.index.isin(df_lungfish_sub2.index.to_list())]

df_sizu.to_csv('sizu_subtitution.txt',sep='\t')

df_sizu

df[df['ortholog pairs'].str.startswith('hu_')]

def gene2Name():
    dic1 = {}
    dic2 = {}
    path='H:/002.singcell/01.liver/10.mutileSpecies/05.SAMap/04.gene2names'
    for i in os.listdir(path):
        if i.endswith('.tsv'):
            sp = i.split('.')[0]
            file =open(path+'/'+i,'r')
            for line in file:
                line = line.strip().split('\t')
                dic1[line[0]] = line[1]
                dic2[sp] = dic1
            file.close()
        else:
            continue
    return dic2
